requirements.txt
streamlit
pandas
numpy
python-dotenv
weaviate-client
sentence-transformers
torch
docling